progetto attualissimo

+++ src/main/java/com/example/ids2425/Repository/AcquirenteRepository.java
@@ -3,5 +3,5 @@ package com.example.ids2425.Repository;
 import com.example.ids2425.Model.Acquirente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AcquirenteRepository extends JpaRepository<Acquirente, Integer> { }
+public interface AcquirenteRepository extends JpaRepository<Acquirente, Long> { }
