+++ src/main/java/com/example/ids2425/Repository/VenditoreRepository.java
@@ -3,4 +3,4 @@ package com.example.ids2425.Repository;
 import com.example.ids2425.Model.Venditore;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VenditoreRepository extends JpaRepository<Venditore, Integer> { }
+public interface VenditoreRepository extends JpaRepository<Venditore, Long> { }
